Move internal implementation to src directory
