+++ lib/open_meteo.dart
@@ -1,17 +1,17 @@
-export 'apis/air_quality.dart';
-export 'apis/climate.dart';
-export 'apis/elevation.dart';
-export 'apis/ensemble.dart';
-export 'apis/flood.dart';
-export 'apis/geocoding.dart';
-export 'apis/historical.dart';
-export 'apis/marine.dart';
-export 'apis/weather.dart';
+export 'src/apis/air_quality.dart';
+export 'src/apis/climate.dart';
+export 'src/apis/elevation.dart';
+export 'src/apis/ensemble.dart';
+export 'src/apis/flood.dart';
+export 'src/apis/geocoding.dart';
+export 'src/apis/historical.dart';
+export 'src/apis/marine.dart';
+export 'src/apis/weather.dart';
 
-export 'enums/air_quality_domain.dart';
-export 'enums/climate_model.dart';
-export 'enums/current.dart';
-export 'enums/daily.dart';
-export 'enums/ensemble_model.dart';
-export 'enums/hourly.dart';
-export 'enums/prefcls.dart';
+export 'src/enums/air_quality_domain.dart';
+export 'src/enums/climate_model.dart';
+export 'src/enums/current.dart';
+export 'src/enums/daily.dart';
+export 'src/enums/ensemble_model.dart';
+export 'src/enums/hourly.dart';
+export 'src/enums/prefcls.dart';
