+++ lib/src/enums/air_quality_domain.dart
