+++ lib/src/enums/climate_model.dart
