+++ lib/src/enums/current.dart
