+++ lib/src/enums/daily.dart
