+++ lib/src/enums/prefcls.dart
