+++ lib/src/exceptions.dart
