+++ lib/src/weather_api_openmeteo_sdk_generated.dart
